\name{prim7}
\alias{prim7}
\docType{data}
\title{prim7 Dataset}
\description{
The data prim7 is a particle physics experiment analyzed
by projection pursuit regression in Friedman and Stuetzle (1981). It has 7 variables on 500
observations. The data set is described in Friedman and Tukey (1974).
}
\format{This data frame contains the following columns:
  \describe{
    \item{X1}{First variable.
    }
    \item{X2}{Second variable.
    }
    \item{X3}{Third variable.
    }
    \item{X4}{Fourth variable.
    }
    \item{X5}{Fifth variable.
    }
    \item{X6}{Sixth variable.
    }
    \item{X7}{Seventh variable.
    }
  }
}
\references{
  Friedman and Tukey (1974),
  A Projection Pursuit Algorithm for Exploratory Data Analysis,
  \emph{IEEE Transactions on Computers} (Volume:C-23,  Issue: 9)

  Friedman, Jerome H.; Stuetzle, Werner (1981),
  Projection pursuit regression.
  \emph{J. Amer. Statist. Assoc.} 76, no. 376, 817--823. 
}

\examples{
data(prim7)
}
\keyword{datasets}
